Update exception? for Expedia::HTTPService::Response
Related to:
https://travelershaven.atlassian.net/browse/HE-9169

The theory for the above linked issue is that somehow this exception?
method was returning true even when the actual Expedia booking request
was successful (it has no EanWsError response) because the content-type
possibly had more characters than exactly 'application/json.'

This causes our he-api side to think there was a booking error when
there actually wasn't because the error category, verbose_message, and
handling never get set since no error actually existed.

As a trial, we are changing this to a regex matcher so that it won't fail
if the strings don't exactly match each other, but rather will only fail
if no part of content-type includes application/json.

More information can be found in the linked JIRA item.